+++ scout/server/extensions/loqus_extension.py
@@ -3,11 +3,15 @@
 * Requires loqusdb version 2.5 or greater.
 * If multiple instances are configured, version will be default's.
 """
+
 import json
 import logging
 import subprocess
 import traceback
 from subprocess import CalledProcessError
+from typing import Dict, Optional
+
+from flask import flash
 
 from scout.exceptions.config import ConfigError
 from scout.utils.scout_requests import get_request_json as api_get
@@ -180,16 +184,9 @@ class LoqusDB:
         }
         return loqus_query
 
-    def get_variant(self, variant_info, loqusdb_id="default"):
-        """Return information for a variant (SNV or SV) from loqusdb
+    def get_variant(self, variant_info: dict, loqusdb_id: str = "default") -> Optional[Dict]:
+        """Return information for a variant (SNV or SV) from loqusdb"""
 
-        Args:
-            variant_info(dict)
-            loqusdb_id(string)
-
-        Returns:
-            loqus_variant(dict)
-        """
         loqus_instance = self.loqusdb_settings.get(loqusdb_id)
         if loqus_instance is None:
             LOG.error(f"Could not find a Loqus instance with id:{loqusdb_id}")
@@ -202,7 +199,7 @@ class LoqusDB:
         return self.get_api_loqus_variant(loqus_instance.get(API_URL), variant_info)
 
     @staticmethod
-    def get_api_loqus_variant(api_url, variant_info):
+    def get_api_loqus_variant(api_url, variant_info) -> dict:
         """get variant data using a Loqus instance available via REST API
 
         SNV/INDELS can be queried in loqus by defining a simple id. For SVs we need to call them
@@ -230,10 +227,15 @@ class LoqusDB:
             search_url = f"{search_url}/?chrom={chrom}&end_chrom={end_chrom}&pos={pos}&end={end}&sv_type={sv_type}"
 
         search_resp = api_get(search_url)
-        if search_resp.get("status_code") != 200:
-            LOG.info(search_resp.get("detail"))
-            return {}
-        return search_resp.get("content")
+
+        if search_resp.get("status_code") == 200:
+            return search_resp.get("content")
+        if search_resp.get("message"):  # An error occurred during the request
+            flash(
+                f"Connection to Loqusdb instance returned error: '{search_resp['message']}'",
+                "warning",
+            )
+        return {}
 
     def get_exec_loqus_variant(self, loqus_instance, variant_info):
         """Get variant data using a local executable instance of Loqus